update user add profile picture
