+++ components/LogoHeader.tsx
@@ -2,7 +2,6 @@
 import Image from "next/image";
 
 export default function LogoHeader() {
-
   return (
     <div className="absolute top-1 left-1/2 -translate-x-1/2 bg-white/10 backdrop-blur-md rounded-full flex flex-col items-center shadow-xl/30 inset-shadow-sm border-white/20 w-11/12 md:w-50 z-30 py-1">
       <div className="flex items-center justify-between w-full px-1">
